models/fish-speech/pyproject.toml
# Generated at 2026-02-02T09:52:12Z from templates/init/pyproject.toml.j2

[project]
name = "ttsdb_fish_speech"
version = "0.1.0"
description = "Fish Speech model for TTSDB"
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    "ttsdb-core>=0.1.4",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "numpy<=1.26.4",
    "transformers>=4.45.2",
    "datasets==2.18.0",
    "lightning>=2.1.0",
    "hydra-core>=1.3.2",
    "tensorboard>=2.14.1",
    "natsort>=8.4.0",
    "einops>=0.7.0",
    "librosa>=0.10.1",
    "rich>=13.5.3",
    "gradio>5.0.0",
    "wandb>=0.15.11",
    "grpcio>=1.58.0",
    "kui>=1.6.0",
    "uvicorn>=0.30.0",
    "loguru>=0.6.0",
    "loralib>=0.1.2",
    "pyrootutils>=1.0.4",
    "resampy>=0.4.3",
    "einx[torch]==0.2.2",
    "zstandard>=0.22.0",
    "pydub",
    "pyaudio",
    "modelscope==1.17.1",
    "opencc-python-reimplemented==0.1.7",
    "silero-vad",
    "ormsgpack",
    "tiktoken>=0.8.0",
    "pydantic==2.9.2",
    "cachetools",
    "descript-audio-codec",
    "descript-audiotools"
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ttsdb_fish_speech"]
include = [
    # Include vendored upstream research code in wheels
    "src/ttsdb_fish_speech/_vendor/**",
]
exclude = [
    # Exclude non-package directories
    "weights/",
    "audio_examples/",
    "space/",
    "tests/",
    "scripts/",
    # Exclude model weight files
    "*.safetensors",
    "*.bin",
    "*.pt",
    "*.pth",
    "*.ckpt",
    "*.h5",
    "*.onnx",
]

[tool.hatch.build.targets.sdist]
exclude = [
    "weights/",
    "audio_examples/",
    "space/",
    ".venv/",
    "dist/",
    "*.safetensors",
    "*.bin",
    "*.pt",
    "*.pth",
    "*.ckpt",
    "*.h5",
    "*.onnx",
]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[tool.hatch.build.targets.wheel.force-include]
"config.yaml" = "ttsdb_fish_speech/config.yaml"
"src/ttsdb_fish_speech/_vendor/source" = "ttsdb_fish_speech/_vendor/source"
